Refresh

+++ vagrant-provision.sh
@@ -5,23 +5,28 @@ export DEBIAN_FRONTEND=noninteractive
 set -e
 
 apt-get update
-apt-get -y upgrade
+apt-get install -y php curl php-mysql mysql-server php-xml php-mbstring php-zip unzip
 
-apt-get install -y php5 curl php5-mysql mysql-server uuid uuid-dev build-essential php-pear
-
-curl -sS https://getcomposer.org/installer | php -- --install-dir=/usr/local/bin
-chmod +x /usr/local/bin/composer.phar
-ln -s /usr/local/bin/composer.phar /usr/local/bin/composer
-
-cd /usr/local/bin
-wget https://phar.phpunit.de/phpunit.phar
-chmod +x phpunit.phar
-ln -s phpunit.phar phpunit
-
-printf "\n" | pecl install uuid
-echo 'extension=uuid.so' > /etc/php5/conf.d/uuid.ini
+if [ ! -f /usr/local/bin/composer ]; then
+  curl -sS https://getcomposer.org/installer | php -- --install-dir=/usr/local/bin
+  chmod +x /usr/local/bin/composer.phar
+  ln -s /usr/local/bin/composer.phar /usr/local/bin/composer
+fi
 
 cd /vagrant
-composer install
+su -c "cd /vagrant && composer install" vagrant
+
+if [ ! -f /etc/mysql/conf.d/vagrant.cnf ]; then
+  mysql -uroot -e "update mysql.user set plugin = 'mysql_native_password' where User='root';"
+
+  tee -a /etc/mysql/conf.d/vagrant.cnf <<EOF
+[mysqld]
+innodb_file_per_table
+slow_query_log = 1
+secure_file_priv=""
+EOF
+
+  service mysql restart
+fi
 
 mysql -uroot -e 'CREATE DATABASE IF NOT EXISTS emphloyer_test'
